refactor: extract some functions
